+++ tests/tinytest/test_funs_string_lazy.R
@@ -212,7 +212,7 @@ expect_equal_lazy(
 
 expect_warning(
   mutate(test, foo = str_extract_all(x2, "[a-z]+", simplify = TRUE)),
-  "will not be used"
+  "doesn't know how to use some arguments"
 )
 
 
@@ -478,7 +478,7 @@ expect_equal_lazy(
 
 expect_warning(
   mutate(test, foo = trimws(x6, which = "right", whitespace = " ")),
-  "will not be used"
+  "doesn't know how to use some arguments"
 )
 
 
@@ -617,7 +617,7 @@ expect_equal_lazy(
 
 expect_warning(
   mutate(test, foo = grepl("e", x1, ignore.case = TRUE)),
-  "will not be used"
+  "doesn't know how to use some arguments"
 )
 
 
@@ -698,13 +698,13 @@ expect_equal_lazy(
 expect_warning(
   mutate(test, foo = str_split(x8, "-", n = 2)) |>
     pull(foo),
-  "will not be used"
+  "doesn't know how to use some arguments"
 )
 
 expect_warning(
   mutate(test, foo = str_split(x8, "-", simplify = TRUE)) |>
     pull(foo),
-  "will not be used"
+  "doesn't know how to use some arguments"
 )
 
 # split_i ---------------------------------------------------------